economy part 1 fix some problems

+++ app/src/main/java/com/example/economysce/Class/LeavingProbability.java
@@ -3,9 +3,9 @@ package com.example.economysce.Class;
 public class LeavingProbability {
     int FromAge;
     int ToAge;
-    int FiredPercent;
-    int ResignPercent;
-    public LeavingProbability(int fromAge, int toAge, int firedPercent, int resignPercent) {
+    double FiredPercent;
+    double ResignPercent;
+    public LeavingProbability(int fromAge, int toAge, double firedPercent, double resignPercent) {
         FromAge = fromAge;
         ToAge = toAge;
         FiredPercent = firedPercent;
@@ -13,6 +13,6 @@ public class LeavingProbability {
     }
     public int getFromAge() { return FromAge; }
     public int getToAge() { return ToAge; }
-    public int getFiredPercent() { return FiredPercent; }
-    public int getResignPercent() { return ResignPercent; }
+    public double getFiredPercent() { return FiredPercent; }
+    public double getResignPercent() { return ResignPercent; }
 }
